healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I0ad10967bc0c45bad7b9dc1f796a3686c1b2cf99
Signed-off-by: Renayura <renayura@proton.me>